Fix CORS server domain list
The CORS require server host url explicitly set, also the port need
be included.

For local deployment, the port need be added to the server url,
while with DEFAULT openshift edge tls termination, the port should
not be added.

So add new TARGET and SSL_ENABLED parameters to identify the
deploy target env and tls status, and update the server list
accordingly.

By default with Helm chart deploy, the value with TARGET and
TLS_ENABLED don't need be set in the Helm chart as default in the
app/core/config.py are set with:

  TARGET: openshift
  TLS_ENABLED: true

Signed-off-by: Wayne Sun <gsun@redhat.com>