Update redis helm chart values
Update to the correct image registry in order to fix
the bug that occured when deploying redis.

Image registry has change from quay.it to docker.io -
explanation - https://github.com/bitnami-labs/sealed-secrets/issues/822

Add image parameters under redis in the Helm chart to allow us to use
a copied image from quay.io to avoid docker.

Fix the replica count to be 1 and not the default which is 3.

Signed-off-by: Ido Burstein <iburstei@redhat.com>